+++ packages/cat-sdk/src/utils/contractPeripheral.ts
@@ -316,6 +316,7 @@ export class CAT20GuardPeripheral {
     }[],
     txInputCount: number,
     txOutputCount: number,
+    deployerAddr: ByteString,
   ): {
     guard: CAT20GuardVariant
     guardState: CAT20GuardConstState
@@ -359,6 +360,7 @@ export class CAT20GuardPeripheral {
 
     // Create guard state based on the selected guard size
     const guardState = CAT20GuardStateLib.createEmptyState(txInputCountMax)
+    guardState.deployerAddr = deployerAddr
 
     // Process token inputs to get token amounts and script hash mapping
     const { tokenAmounts, tokenScriptHashes, tokenScriptIndexes, guardTokenTypes } = this.processTokenInputs(
@@ -408,7 +410,8 @@ export class CAT20GuardPeripheral {
     tokenInputs: {
       token: UTXO
       inputIndex: number
-    }[]
+    }[],
+    deployerAddr: ByteString,
   ): {
     guard: CAT20GuardVariant
     guardState: CAT20GuardConstState
@@ -436,6 +439,7 @@ export class CAT20GuardPeripheral {
 
     // Create guard state based on the selected guard size
     const guardState = CAT20GuardStateLib.createEmptyState(txInputCountMax)
+    guardState.deployerAddr = deployerAddr
 
     // Process token inputs to get token amounts and script hash mapping
     const { tokenAmounts, tokenScriptHashes, tokenScriptIndexes } = this.processTokenInputs(
@@ -682,6 +686,7 @@ export class CAT721GuardPeripheral {
     receivers: ByteString[],
     txInputCount: number,
     txOutputCount: number,
+    deployerAddr: ByteString,
   ): {
     guard: CAT721GuardVariant
     guardState: CAT721GuardConstState,
@@ -713,6 +718,7 @@ export class CAT721GuardPeripheral {
 
     // Create guard state to get the initial nftScriptIndexes
     const guardState = CAT721GuardStateLib.createEmptyState(txInputCountMax)
+    guardState.deployerAddr = deployerAddr
 
     // Process NFT inputs to get script hash mapping
     const { nftScriptHashes, nftScriptIndexes, guardCollectionTypes } = this.processNftInputs(
@@ -752,7 +758,8 @@ export class CAT721GuardPeripheral {
     nftInputs: {
       nft: UTXO,
       inputIndex: number
-    }[]
+    }[],
+    deployerAddr: ByteString,
   ): {
     guard: CAT721GuardVariant
     guardState: CAT721GuardConstState,
@@ -779,6 +786,7 @@ export class CAT721GuardPeripheral {
 
     // Create guard state to get the initial nftScriptIndexes
     const guardState = CAT721GuardStateLib.createEmptyState(txInputCountMax)
+    guardState.deployerAddr = deployerAddr
 
     // Process NFT inputs to get script hash mapping
     const { nftScriptHashes, nftScriptIndexes, guardCollectionTypes } = this.processNftInputs(