+++ packages/scrypt-ts-cli-opcat/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @opcat-labs/cli-opcat
 
+## 3.2.0
+
+### Patch Changes
+
+- @opcat-labs/scrypt-ts-transpiler-opcat@3.2.0
+
 ## 3.1.0
 
 ### Minor Changes